+++ packages/apollo-angular/src/mutation.ts
@@ -1,8 +1,6 @@
 import {Injectable} from '@angular/core';
 import {DocumentNode} from 'graphql';
 import {TypedDocumentNode} from '@graphql-typed-document-node/core';
-import {Observable} from 'rxjs';
-import {FetchResult} from '@apollo/client/core';
 
 import {Apollo} from './apollo';
 import {MutationOptionsAlone, EmptyObject} from './types';
@@ -14,10 +12,7 @@ export class Mutation<T = {}, V = EmptyObject> {
 
   constructor(protected apollo: Apollo) {}
 
-  public mutate(
-    variables?: V,
-    options?: MutationOptionsAlone<T, V>,
-  ): Observable<FetchResult<T>> {
+  public mutate(variables?: V, options?: MutationOptionsAlone<T, V>) {
     return this.apollo.use(this.client).mutate<T, V>({
       ...options,
       variables,