+++ packages/apollo-angular/src/types.ts
@@ -4,6 +4,7 @@ import {
   MutationOptions as CoreMutationOptions,
   SubscriptionOptions as CoreSubscriptionOptions,
   ApolloClientOptions,
+  FetchResult,
 } from '@apollo/client/core';
 import {ExecutionResult} from 'graphql';
 
@@ -15,6 +16,10 @@ export interface ExtraSubscriptionOptions {
   useZone?: boolean;
 }
 
+export interface MutationResult<TData = any> extends FetchResult<TData> {
+  loading: boolean;
+}
+
 export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
 
 export interface WatchQueryOptionsAlone<TVariables = EmptyObject, TData = any>
@@ -24,10 +29,7 @@ export interface QueryOptionsAlone<TVariables = EmptyObject, TData = any>
   extends Omit<CoreQueryOptions<TVariables, TData>, 'query' | 'variables'> {}
 
 export interface MutationOptionsAlone<TData = EmptyObject, TVariables = any>
-  extends Omit<
-    CoreMutationOptions<TData, TVariables>,
-    'mutation' | 'variables'
-  > {}
+  extends Omit<MutationOptions<TData, TVariables>, 'mutation' | 'variables'> {}
 
 export interface SubscriptionOptionsAlone<TVariables = EmptyObject, TData = any>
   extends Omit<
@@ -46,6 +48,17 @@ export interface WatchQueryOptions<TVariables = EmptyObject, TData = any>
   useInitialLoading?: boolean;
 }
 
+export interface MutationOptions<TVariables = EmptyObject, TData = any>
+  extends CoreMutationOptions<TVariables, TData> {
+  /**
+   * Observable starts with `{ loading: true }`.
+   * There's a big chance the next major version will enable that by default.
+   *
+   * Disabled by default
+   */
+  useMutationLoading?: boolean;
+}
+
 export interface SubscriptionResult<TData> extends ExecutionResult {
   data?: TData;
 }
@@ -60,4 +73,10 @@ export type Flags = {
    * Disabled by default
    */
   useInitialLoading?: boolean;
+  /**
+   * Observable starts with `{ loading: true }`.
+   *
+   * Disabled by default
+   */
+  useMutationLoading?: boolean;
 };