+++ packages/apollo-angular/CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fix `useInitialLoading` [`750429c`](https://github.com/kamilkisiela/apollo-angular/commit/750429cbe0aaad07a3fdc2e2ef046fe8be4aad28)
 - Typed `gql` and `graphql` tags - both accept two generic types and support `TypedDocumentNode` [`9a8ea5f`](https://github.com/kamilkisiela/apollo-angular/commit/9a8ea5f229cf7937d74332092cb3eba40828b7b1)
 - Add `useMutationLoading` flag [`bc223fe`](https://github.com/kamilkisiela/apollo-angular/commit/bc223fe6487edd35c56ad908e4739580ce69f056)
+- Fix type inference for Mutations [#1659](https://github.com/kamilkisiela/apollo-angular/pull/1659)
 
 ### v2.4.0