+++ packages/apollo-angular/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ### vNext
 
+### v2.5.0
+
+- Support Angular 12
+
 ### v2.4.0
 
 - Fix: Use let in Apollo constructor due to firefox bug [#1632](https://github.com/kamilkisiela/apollo-angular/pull/1632)